requirements.txt
# Gateway + core + local tooling (Python 3.10+)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
redis>=5.0.0
asyncpg>=0.29.0
python-multipart>=0.0.6
sse-starlette>=1.8.0
httpx>=0.26.0
jinja2>=3.1.0
numpy>=1.26.0
langgraph>=0.2.0
langchain-core>=0.2.0
PyYAML>=6.0.0

# Optional: EMBEDDING_BACKEND=bge (BAAI/bge-m3)
sentence-transformers>=2.2.0
torch>=2.0.0

# Optional: MinIO client (Phase C object storage)
minio>=7.2.0
